fix(platform-wallet-storage): secrets hardening — atomic reprotect, scheme-0 zeroize, version width (#3968 review)
H1: SecretStore::reprotect did a non-atomic get→set; a concurrent set/delete between them could let a transform built on stale bytes clobber a newer value. The File arm now runs the read→rewrap→write under the store's single lock (new EncryptedFileStore::reprotect_bytes). The Os arm is a per-item keyring with no transaction, so its non-atomicity is now documented as a residual.

H3: the scheme-0 (unprotected) wrap left a cleartext plaintext copy in the envelope's Vec unzeroized; it's now wiped after encoding (the returned SecretBytes is the only retained copy). L6: UnsupportedEnvelopeVersion stored the version as u8, truncating Envelope.version (u32) — now u32 so a >255 version isn't aliased in diagnostics (with a regression test).

L5: narrowed the error-module non-leakage doc — the Io variant intentionally carries its non-secret OS source; every other variant is source-free. L7: documented the vault's read-side duplicate-JSON-key collapse (serde_json last-wins), benign because every entry is AEAD-bound to (wallet_id, label).


<sub>🤖 Co-authored by [Claudius the Magnificent](https://github.com/lklimek/claudius) AI Agent</sub>

+++ packages/rs-platform-wallet-storage/src/secrets/error.rs
@@ -1,10 +1,15 @@
 //! Secret-store error taxonomy and its `keyring_core::Error` projection.
 //!
 //! Variants carry only non-secret diagnostics (POSIX mode bits, header
-//! version, vault path) — never a secret byte, passphrase, plaintext, or
-//! stringified source (CWE-209/CWE-532). The public, fully-typed path is
-//! the [`SecretStore`](crate::secrets::SecretStore) API; the SPI
-//! projection into `keyring_core::Error` is lossy (see the [`From`] impl).
+//! version, vault path) — never a secret byte, passphrase, or plaintext
+//! (CWE-209/CWE-532). The single carried source is the [`Io`] variant's
+//! OS error (an errno plus the non-secret caller-supplied path); every
+//! other variant is source-free so a crypto/format failure can't stringify
+//! a secret. The public, fully-typed path is the
+//! [`SecretStore`](crate::secrets::SecretStore) API; the SPI projection into
+//! `keyring_core::Error` is lossy (see the [`From`] impl).
+//!
+//! [`Io`]: SecretStoreError::Io
 
 use std::path::Path;
 
@@ -86,9 +91,10 @@ pub enum SecretStoreError {
     /// [`VersionUnsupported`]: SecretStoreError::VersionUnsupported
     #[error("unsupported secret envelope version {found}")]
     UnsupportedEnvelopeVersion {
-        /// The envelope `version` byte read from the (unauthenticated)
-        /// header.
-        found: u8,
+        /// The full `version` field read from the (unauthenticated)
+        /// envelope header. `u32` to match `Envelope.version` — a truncating
+        /// `u8` would alias distinct out-of-range versions in diagnostics.
+        found: u32,
     },
 
     /// The vault file was malformed (bad magic, truncated header, bad
